Move pipeline features to DSL (#39)
This PR refactors Pipeline and Processors, completely. We’re renaming Processors to Extensions and making them even more powerful than before. Extensions now have the ability to add to the pipeline DSL. It’s such a great pattern that we use Extensions to implement native features.

This refactoring also allows us to add group support. This is a breaking change and a major version bump is required.